third_party/WebKit/LayoutTests/platform/linux/external/wpt/css/css-tables/table-model-fixup-2-expected.txt

This is a testharness.js-based test.
PASS Replaced elements inside a table cannot be table-row and are considered block -- input elements
PASS Replaced elements inside a table cannot be table-row and are considered block -- img elements
PASS Replaced elements inside a table cannot be table-column and are considered inline -- input elements
PASS Replaced elements inside a table cannot be table-column and are considered inline -- img elements
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- input elements
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- img elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=button elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=file elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- img elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=text elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=button elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=file elements
FAIL Replaced elements outside a table cannot be table and are considered block -- img elements assert_not_equals: got disallowed value 2197
FAIL Replaced elements outside a table cannot be table-row and are considered block -- input=text elements assert_not_equals: got disallowed value 2299
FAIL Replaced elements outside a table cannot be table-row and are considered block -- input=button elements assert_not_equals: got disallowed value 2423
PASS Replaced elements outside a table cannot be table-row and are considered block -- input=file elements
FAIL Replaced elements outside a table cannot be table-row and are considered block -- img elements assert_not_equals: got disallowed value 2695
FAIL Replaced elements outside a table cannot be table-row-group and are considered block -- input=text elements assert_not_equals: got disallowed value 2797
FAIL Replaced elements outside a table cannot be table-row-group and are considered block -- input=button elements assert_not_equals: got disallowed value 2921
PASS Replaced elements outside a table cannot be table-row-group and are considered block -- input=file elements
FAIL Replaced elements outside a table cannot be table-row-group and are considered block -- img elements assert_not_equals: got disallowed value 3193
PASS Replaced elements outside a table cannot be table-column and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-column and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-column and are considered inline -- input=file elements assert_equals: expected 3567 but got 3545
PASS Replaced elements outside a table cannot be table-column and are considered inline -- img elements
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-cell and are considered inline -- input=file elements assert_equals: expected 4065 but got 4043
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- img elements
Harness: the test ran to completion.
